+++ js/glotdict-validation.js
@@ -53,7 +53,10 @@ function gd_search_glossary_on_translation( e, selector ) {
 			original_index = PLURAL; // For locales where nplural === 1.
 		}
 
-		translations.each( ( i, translation ) => {
+		translations.each((i, translation) => {
+			// PSS issue issue #120 check for URL
+			var translatedText = translation.value
+
 			const glossary_words = jQuery( '.glossary-word', originals[original_index] ).map( function() {return this.textContent;} ).get();
 			const words_with_warning = [];
 			jQuery( '.glossary-word', originals[original_index] ).each( ( j, glossary_element ) => {
@@ -90,7 +93,17 @@ function gd_search_glossary_on_translation( e, selector ) {
 						message = 'The translation does not contain any of the suggested translations';
 					}
 					const form = translations.length > 1 ? ( original_index === SINGULAR ? ' for singular' : ' for plural' ) : '';
-					jQuery( '.textareas', $editor ).prepend( gd_get_warning( `${message} (${reset}) for the term “<i>${term}</i>“ ${count}${form}.`, discard ) );
+					// PSS issue issue #120 check for URL
+					let is_within_URL = gd_check_for_URL(glossary_word, translatedText)
+					// If it is not part of an URL give the warning for it
+					if (is_within_URL == false) {
+						jQuery('.textareas', $editor).prepend(gd_get_warning(`${message} (${reset}) for the term “<i>${term}</i>“ ${count}${form}.`, discard));
+					}
+					else {
+						// We need not to mark the current as wrong
+						howmany--
+						return howmany;
+					}
 				}
 			} );
 			original_index = PLURAL;