+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Bugfix: The button to hide sticky header was no longer shown
 * Enhancement: Buttons do now show Tooltip
+* Enhancement: Add function to check if glossary word is within a URL
 
 
 # 2.1.4